Added a keyword to parse-decimal-number.
I was looking for a way of parsing the first character from the number
portion of a number.  When dealing with signed numbers, this was difficult
without first knowing if there was a sign or not.  I've added a keyword
to allow parsing of the number with :START and :END keywords free of any
sign the string may posess.